Configured post sending with attachments

+++ TGHub.Application/Interfaces/ITgHubTelegramBotClient.cs
@@ -5,5 +5,5 @@ namespace TGHub.Application.Interfaces;
 public interface ITgHubTelegramBotClient
 {
     Task CreateOrUpdateChannelFromTg(long channelTgId);
-    Task<long> SendPostAsync(Post post);
+    Task<int> SendPostAsync(Post post);
 }
